moves the patches into the RedmineOpenpgp namespace
